Move linkedList to utils
