+++ utils/linkedList/index.js
