+++ utils/linkedList/linkedListNode.js
